sui_metrics_push_client
Function start_metrics_push_task

pub fn start_metrics_push_task(
    push_interval_seconds: Option<u64>,
    push_url: String,
    metrics_key_pair: NetworkKeyPair,
    registry: RegistryService,
)

Starts a task to periodically push metrics to a configured endpoint if a metrics push endpoint is configured.